pyproject.toml
[project]
name = "llm-eval"
version = "0.1.0"
description = "LLM-based machine translation evaluation toolkit using Azure OpenAI instead of COMET models"
authors = [ { name = "Your Name", email = "you@example.com" } ]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "httpx>=0.27.0",
  "azure-ai-inference>=1.0.0b5",  # retained for optional services.ai usage (may be removed later)
  "openai>=1.30.0",
  "python-dotenv>=1.0.0",
  "tqdm>=4.66.0",
  "numpy>=1.26.0",
  "scikit-learn>=1.4.0",
  "typer>=0.12.3",
  "pydantic>=2.7.0"
]

[project.optional-dependencies]
embed = ["sentence-transformers>=2.7.0"]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=5.0.0"
]

[project.scripts]
llm-eval = "llm_eval.cli:app"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
extend-select = []

